+++ library/src/test/java/com/chuckerteam/chucker/api/ChuckerInterceptorTest.kt
@@ -18,6 +18,7 @@ import okhttp3.RequestBody
 import okhttp3.RequestBody.Companion.toRequestBody
 import okhttp3.mockwebserver.MockResponse
 import okhttp3.mockwebserver.MockWebServer
+import okhttp3.mockwebserver.SocketPolicy
 import okio.Buffer
 import okio.BufferedSink
 import okio.ByteString
@@ -496,6 +497,31 @@ internal class ChuckerInterceptorTest {
             """.trimIndent()
         )
         assertThat(transaction.requestPayloadSize).isEqualTo(request.body!!.contentLength())
+    }
+
+    @ParameterizedTest
+    @EnumSource(value = ClientFactory::class)
+    fun requestHeaders_areRedacted_whenServerFails(factory: ClientFactory) {
+        val chuckerInterceptor = ChuckerInterceptorDelegate(
+            maxContentLength = SEGMENT_SIZE,
+            headersToRedact = setOf("Header-To-Redact"),
+            cacheDirectoryProvider = { tempDir },
+        )
+        val client = factory.create(chuckerInterceptor)
+
+        val request = Request.Builder().url(serverUrl).header("Header-To-Redact", "Hello").build()
+        server.enqueue(MockResponse().setSocketPolicy(SocketPolicy.DISCONNECT_AT_START))
+        runCatching { client.newCall(request).execute() }
+
+        val transaction = chuckerInterceptor.expectTransaction()
+        assertThat(transaction.requestHeaders).contains(
+            """
+            |  {
+            |    "name": "Header-To-Redact",
+            |    "value": "**"
+            |  }
+            """.trimMargin()
+        )
     }
 
     @ParameterizedTest